cloud-function/requirements.txt
mcrcon
dialogflow
google-api-python-client
